extend token when already logged in and try to login again
